CSEM and CSSEM with Binomial Model

Compute the CSEM and CSSEM using binomial model

csem.binomial(ni, ct = NULL)

Arguments

ni

a numeric value indicating the number of items

ct

A data frame or matrix containing conversion table with the first column as raw scores and the second column as scale scores

Value

a list containing raw scores and corresponding CSEM using bimonial model, and CSSEM if the conversion table is provided

Details

References

Examples

csem.binomial(40)
#> $x
#>  [1]  0  1  2  3  4  5  6  7  8  9 10 11 12 13 14 15 16 17 18 19 20 21 22 23 24
#> [26] 25 26 27 28 29 30 31 32 33 34 35 36 37 38 39 40
#> 
#> $csem
#>  [1] 0.000000 1.000000 1.395965 1.687055 1.921538 2.118296 2.287087 2.433737
#>  [9] 2.562050 2.674667 2.773501 2.859980 2.935198 3.000000 3.055050 3.100868
#> [17] 3.137858 3.166329 3.186510 3.198557 3.202563 3.198557 3.186510 3.166329
#> [25] 3.137858 3.100868 3.055050 3.000000 2.935198 2.859980 2.773501 2.674667
#> [33] 2.562050 2.433737 2.287087 2.118296 1.921538 1.687055 1.395965 1.000000
#> [41] 0.000000
#> 
csem.binomial(40, ct.u)
#> $x
#>  [1]  0  1  2  3  4  5  6  7  8  9 10 11 12 13 14 15 16 17 18 19 20 21 22 23 24
#> [26] 25 26 27 28 29 30 31 32 33 34 35 36 37 38 39 40
#> 
#> $csem
#>  [1] 0.000000 1.000000 1.395965 1.687055 1.921538 2.118296 2.287087 2.433737
#>  [9] 2.562050 2.674667 2.773501 2.859980 2.935198 3.000000 3.055050 3.100868
#> [17] 3.137858 3.166329 3.186510 3.198557 3.202563 3.198557 3.186510 3.166329
#> [25] 3.137858 3.100868 3.055050 3.000000 2.935198 2.859980 2.773501 2.674667
#> [33] 2.562050 2.433737 2.287087 2.118296 1.921538 1.687055 1.395965 1.000000
#> [41] 0.000000
#> 
#> $cssem
#>  [1] 0.000000 1.177056 1.899213 2.283665 2.437251 2.453570 2.397404 2.305917
#>  [9] 2.200722 2.095503 1.997789 1.910156 1.832986 1.766998 1.713704 1.674081
#> [17] 1.647028 1.628875 1.614108 1.596773 1.572053 1.537792 1.495812 1.452585
#> [25] 1.418499 1.405001 1.419997 1.463925 1.529442 1.605029 1.679935 1.747840
#> [33] 1.807565 1.859475 1.897988 1.905824 1.859935 1.752578 1.606934 1.407919
#> [41] 0.000000
#>