CSEM with Compound Binomial Model

Compute the CSEM using compound binomial model

csem.compound.binomial(x, s)

Arguments

x

A data frame or matrix containing data with rows as subjects and items as columns

s

A vector containing number of items for each strat

Value

a list containing raw scores and corresponding CSEM using compound bimonial model

Details

References

Examples

data(data.m)
csem.compound.binomial(data.m, c(13, 12, 6))
#> $x
#>    [1] 15 22 19 27 22 11 20 20 23 11 23 18 27 23 19 17 20 19 18 21 23 16 22 20
#>   [25] 23 24 24 19 25 26 20 20 17 25 17 23 15 25 18 23 16 25 19 22 21  7 26 23
#>   [49] 17 23 23 22 11 15 13 25 20 17 27 17 28 20 28 18 21 20 19 25 15 25 16 13
#>   [73] 23 22 20 23 27 17 22 19 25 19 22 20 19 20 25 24 14 16 18 15 18 21 23 18
#>   [97] 24 19 20 14 24 25 16 20 25 28 21 23 21 18 24 25 25 18 13 21 28 27 16 20
#>  [121] 16 23 22 21 22 22 20 22 21 16 23 29 27 21 12 16 22 29 24 23 23 18 17 22
#>  [145] 25 26 27 19 19 23 20 19 29 25 30 17 11 25 26 23 23 22 22 28 13 12 12 22
#>  [169] 20 21 17 20 19  9 20 26 26 17 25 10 26 23 24 26 27 20 20 24 24 20 22 22
#>  [193] 16 20 17 20 23 20 24 27 21 21 15 20 20 23 15 24 21 13 24  8 19 24 24 22
#>  [217] 27 19 22  4 15 27 22 11 19 21 29 23 26 22 28 17 19 22 25 20 16 23 14 17
#>  [241] 24 21 19 20 16 22 18 21 22 18 23 25 22 19 22 18 15 18  9 15 26 17 25 23
#>  [265] 23 21 24 30 12 27 19 24 23 23 24 20 17 27 16 28 11 17 19 27 24 23 24 21
#>  [289] 19 22 18 20 23 23 25 18 14 24 26 11 17 23 20 25 25 21 22 17 10 26 22 17
#>  [313] 14 20  9 16 29 20 26 15 23 19 27 26 20 17 21 15 26 18 25 28 24 23 23 24
#>  [337] 23 22 27 17 26 25 17  9 18 24 25 23 25 12 23 20 27 20 13 26 23 17 19 26
#>  [361] 19  5 15 24 16 15 26 15 21 25 21 23 22 23 22 24 15 27 20 24 24 23 25 22
#>  [385] 10 20 24 17 23 27 24 31 28 25 15 20 14 12 29 18 22 14 25 24 22 19 20 27
#>  [409] 23 16 23 14 24 16 17 17 20 22 17 18 18 21 17 20 22 25 27 20 27 20 18 24
#>  [433] 18 19 20 15 29 24 13 22 27 15 22 20 17 24 18 25 21 15 21 13 20 24  5 26
#>  [457] 22 20 21 21 25 30 22 18 17 18 12 18 21 27 12 20 22 10 16 22 21 23 22 24
#>  [481] 17 25 10 23 23 12 25 20 24 20 21 28 23 25 15 25 19 22 18 23 24 27 20 20
#>  [505]  5 15 19 19 28 23 23 24 22 11 12 20 18 16 21 18 23 20 20 23 21 17 22 18
#>  [529] 20 23 21 20 24 24 13 17 22 20 24 26 19 24 22 20 17 21 25 27 13 14 25 15
#>  [553] 19 22 14 25 17 25 26 13 26 15 18 10 24 26 21 13 25 22  8 18 23 18 21 26
#>  [577]  9 25 27 16 21 19 22 19 25 23 21 22 24 22 20 11 22 15 26 19 20 21 26 23
#>  [601] 25 15 20 26 20 18 13 24 18 25 24 26 17 24 12 21 22 24 17 18 19 17 25 25
#>  [625] 25 25 26 25 29 23 23 24 20 15 23 27 24 26 24 24 24 29 25 25 27 20 23 18
#>  [649] 21 26 11 25 16 22 25 21 17 25 22 19 27 26 16 22 16 17 22 14 27 23 21 28
#>  [673] 26 19 13 24 27 25 16 18 17 17 20 18 25 23 30 20 15 14 23 19 23 25 23 24
#>  [697] 19  8 22 20 20 18 22 16 14 23 17 11 25 25  8 19 16 11 24 22 18 24 25 19
#>  [721] 17  7 22 24 21 20 26 24 22 27 26 18 23 23 25 25 14 23 22 24 18 25 20 19
#>  [745] 25 23 13 21 13 20 15 21 23 23 18 19 19 19 23 20 22 14 24 19 17 23 24 15
#>  [769] 14 20 23 21 22 25 25 20 21 19 21 22 24 14 16 21 18 14 26 11 22 19 27 13
#>  [793] 14 18 23 16 18 19 22 28 24 21 21 19 21 20 14 26 28 21 19 22 21 13 25 26
#>  [817] 23 23 14 12 24 23 19 25 28 23 24 20 20 24 18 16 20 15 11 21 20 19 15 23
#>  [841] 24 19 20 20 20 15 21 23 26 20 24 23 23 17 25 25 21 24 23 17 27 18 24 24
#>  [865] 23 28 23 14 22 25 20 21 20 19 22 19 29 22 15 27 25 26 28 20 18 23 19 18
#>  [889] 21 14 18 18 17 19 11 25 27 21 24  9 14 17 24 23 16 24 14 22 21 26 26 18
#>  [913] 21 15 20 25 26 22 19 25 21 20 14 21 24 30 23 25 25 15 20 16 23 19 22 16
#>  [937]  8 23 28 22 24 27 19 28 23 22 27 21 17 19 14 27 23 14 16 28  6 28 25 25
#>  [961] 20 16 22 20 16 25 23 20 23 24 14 17 26 22 27 20 16 24 10 24 26 25 24 24
#>  [985] 21 22 23 17 20 20 18 17 20 21 27 25 24 13 26 20 18 16 23 18 19 23 24 27
#> [1009] 24 26 18 24 25 26 24 25 28 26 21 20 22 21 26 19 23 26 18 23 22 24 22 27
#> [1033] 23 23 13 24 21 17 24 23 20 23 21 19 26 16 21 26 23 25 19 20 26 28 22 18
#> [1057] 25  6 14 26 22 22 21 25 20 24 23 10  8 24 25 27 22 27 25 24 13 19 11 21
#> [1081] 20 15 22 26 25 22 15  9 19 24 26 24 14 19 27 21 23 20 16 19 20 25 19 10
#> [1105] 23 18 27 22 14 25 22 16 23 24 27 21 17 24 19 22 22 21 11 24 15 23 22 20
#> [1129] 26 22  9 27 18 20 21 18 28 20 17 23 27 22 22 17 17 24 17 23  8 20 15 21
#> [1153] 21 23 21 15 13 11 30 16 25 19 10 20 20 10 19 19 23 28 23 17 20 24 22 15
#> [1177] 19 24 26 12 11  7 25 11 25 20 19 28 28 18 22 19 18 23 14 25 16 18 17 23
#> [1201] 20 22 22 22 21 27 16 23 19 14 30 26 22 21 24 24 11 20 19 21 14 27 14 15
#> [1225] 16 27 24 15 17 22 15 19 23 20 29 28 26 17 26 24 16 19 21 17 26 14 19 15
#> [1249] 25 21 25 22 22 15 23 20 20 17 26 23 14 22 11 15 21 25 16 22 14 23 25 19
#> [1273] 22 26 22 26 18 26 18 22 15  9 13 21 26 19 17 16 23 27 24 20 21 20 18 27
#> [1297] 22 23 26 19 12 23 21 16 20 18 15 25 25 20 20 21 18 23 25 30 17 13 18 21
#> [1321] 16 14 27 31 27 12 26 20 15 26 18 23 22 21 20 22 23  7 19 16 19 20 24 23
#> [1345] 24 20 25 28 19 20 15 19 17 18 19 11 18 25 22 16 27 25 19 14 20 24 21 20
#> [1369] 23 21 25 16 28 20 24 10 20  9 30 19 23 26 23 22 23 15 17 26 18 27 25 14
#> [1393] 20 25 29 13 22 21 18 23 18 12 25 24 23 16 15 21 20 13 19 23 20 23 20 22
#> [1417] 27 21 15 24 25 27 23 24 19 28 12 26 25 19 14 23 18 22 24 26 25 25 22 24
#> [1441] 22 29 12 17 19 24 24 16 28 20 20 13 25 19 23 23 20 17 18 19 22 16 27 25
#> [1465] 21 21 14 17 24 26 24 19 22 22 23 20 19 20 27 13 13 17 21 18 15 23 27 19
#> [1489] 15 22 13 12 22 23 19 19 18 23 19 26 22 23 20 17 29 25 21 15 22 24 16 28
#> [1513] 26 10 15 19 17 23 26 23 15 28 25 11 21 22 15 19 19 26 26 15 25 22 17 23
#> [1537] 24 25 13 23 24 22 21 30 15  5 26 19 27 16 16 14 24 17 25 22 13 20 18 23
#> [1561] 17 24 18 16 23 23 18 22 18 22 19 17 17 23 25  9 24 19 27 16 18 12 20 21
#> [1585] 15 20 20 18 20 25 25 15 23 13 24 16 21 23 21 24 19 18 20 21 24 14 13 18
#> [1609] 25 21 25 25 10 23 29 24 23 22 20 25 25 14 28 24 22 18 18 23 23 26 18 12
#> [1633] 29 29 19 16 24 20 22 26 21 15 26 13 20 25 19 19 21 20 23 14 26 26 21 19
#> [1657] 16 19 24 22 28 16 22 26 28 20 20 29 27 23 27 22 15 21 21 23 29 25 12 19
#> [1681] 23 17 20 20 30 18 22 26 14 19 22 27 21 20 24 25 22 22 11 21 11 21 18 19
#> [1705] 22 27 20 15 15 21 21 17 21 22 25 15 22 15 16 22 17 23 24 20 25 17 21 25
#> [1729] 23 24 19 27 23 26 25 16  5 26 22 19 17 24 24 22 21 19 23 19 25 21 26 15
#> [1753] 13 23 29 19 24 27 22 20 22 17 25 15 21 22 24 20 24 29 14 23 16 25 22 18
#> [1777] 17 25 27 12 27 17 22 26 28 19 24 21 22 20 23 23 18 13 19 25 24 21 22 31
#> [1801] 27 18  7 24 28 23 23 23 22 25 28 23 12 23 18 21 23  9 26 21 12 17 25 20
#> [1825] 22 20 26 22 15 20 17 24 18 25 20 19 19 20 22 22 23 25 14 17 26 23 24 28
#> [1849]  9 17 24 19 16 22 23 27 22 28 26 18 13 19 23 19 24 15 20 11 25 21 17 23
#> [1873] 27 16 19 17 23 24 21 17 20  6 24 25 16 24 23 23 25 25 16 26 22 21 15 12
#> [1897] 20 25 20 25 25 22 21 24 25 24 20 17 15 13 15 25 24 23 29 16 24 26 12 28
#> [1921] 23 18 23 26 27 18 10 20 16 13 21 22  8 16 20 28 11 23 27 17 20 26 23 17
#> [1945] 23 15 21 23 26 21 22 28 20 17 16 23 22 21 22 19 24 25 10 22 25 21 26 17
#> [1969] 17 17 23 16 18 22 24 19 23 11 18 22 21 22 19 24 22 20 10 27 21 21 18 19
#> [1993] 14 16 20 22 22 22 28 26 21 26 24 23 23 20 18 18 24 26 14 24 20 18 21 17
#> [2017] 22 26 19 20 19 16 24 22 20 18 27 23 23 28 25 13 27 12 20 22 22 24 17 16
#> [2041] 20 28 21 17 22 24 17 17 27 19 21 18 23 23 25 25 26 27 21 19 25 26 19 26
#> [2065] 20 23  8 24 25 22 24 29 19 25 23 19 24 19 22 20 14 20 17 20 25 20 15 14
#> [2089] 24 19 20 23 17 18 19 17 23 23 18 24 23 24 16 19 26 24 24 10 26 17 20 21
#> [2113] 18 18 24 20 15 22 11 17 21 25 26 20 28 17 26 20 11 25 23 25 28 26 22  9
#> [2137] 22 22 17 16 22 19  8 22 17 26 12 23 19  8 20 12 21 20 21 16 19 17 27 23
#> [2161] 22 20 20 18 11 14 25 20 24 22 16 16 16 19 21 25 14 19 27 25 24 22 28 16
#> [2185] 21 23 18 20 28 18 25 18 20 22 22 22 28 22 10 23 16 20 21 22 26 26 21 21
#> [2209] 26 18 25 25 21 27 25 24 22 16 21 17 22 23 20 24 23 21 15 27 23 24 16 12
#> [2233] 22 20 18 21 19 19 22 22 22 16 22 23 24 19 31 27 27 21 26 28 25 22 13 20
#> [2257] 27 25 21 18 12 23 23 21 20 22 21 29 22 17  9 19 19 29 23 22 10 26 22 22
#> [2281] 14 24 13 23 24 26 25 23 22 20 24 17 26 16 24 25 22 27 25 23 21 27 17 20
#> [2305] 20 20 27 27 20  9 18 21 28 12 22 23 28 18 20 19 18 28 21 22 25 19 19 23
#> [2329] 14 21 25 25 26 24 25 29 15 12 24 23 16 17 21 17 19 11 19 21 23 27 25 11
#> [2353] 18 12 20 20 30 25 28 23 26 21 21 24 22 20 25 21 24 22 24 16 17 22  9 22
#> [2377] 19 14 22 23 21 27 24 21 22 21 21 17 14 25 20 24 20 22 21 28 22 17 27 19
#> [2401] 21 22 18 17 23 21 18 27 18 22 29 27 25 17 21 23 10 21 23 23 20 15 19 11
#> [2425] 25 15 12 22 19 22 24 27 14 23 27 18  9 11 23 23 23 15 24 16 23 19 22 24
#> [2449] 25 13 23 25 22 18 15 27 18 20 27 23 22 18 18 24 24 27 26 17 19 29 23 16
#> [2473] 18 21 24 26 24 19 25 16 16 20 23 17 22 22 15 28 22 19 24 18 20 15 24 23
#> [2497] 18 19 18 23 16 24 23 22 22 22 21 18 28 13 21 14 24 17 23 20 22 15 21 14
#> [2521] 18 13 24 22 26 26 13 22 28 26 16 11 18 16 14 26 26 21 22 22 20 21 12 22
#> [2545] 27 20 28 23 11 19 22 14 23 21 23 21 26 24 28 21 24 28 27 18 25 22 12  8
#> [2569] 23 21 20 18 18 19 23 23 24 24 16 13 26 18 22  8 19 13 22 24 18 12 27 16
#> [2593] 13 14 18 23 20 14 25 17 22 28 23 24 12 25 24 24 23 15 16 24  8 18 29 27
#> [2617] 24 18  8 22 27 26 23 21 21 23 23 25 14 23 16 20 17 13 22 26 25 21 17 20
#> [2641] 18 25 18 18 21 20 28 21 15 26 21 25 21 26 19 18 16 23 21 18 19 18 17 23
#> [2665] 18 18 21 20 23 26 21 23 26 12 26 18 16 17 16 17 23 20 10 24 19 24 25 19
#> [2689] 14 11 25 19 27 19 22 22 13 21 19 18 13 26 20 11 27 11 20 18 25 17 25 23
#> [2713] 16 22 23 20 15 16 22 23 19 24 16 26 11 20 13 21 24 20 23 28 20 16 20 15
#> [2737] 26 23 19 18 21 27 27 23 22 23 22 24 19 23 25 17 21 24 12 17  6 27 27 17
#> [2761] 20 24 19 24 22 17 26 20 23 27 24 21 20 26 21 13 17 23 17 17 21 18 14 13
#> [2785] 17 24 23 24 24  8 17 19 17 14 24 18 21 22 19 20 21 23 24 19 25 25 20 20
#> [2809] 30 14 30 13 14 22 14 19 22 23 10 19 21 21 14 25 13 21 15 16 23 20 28 27
#> [2833] 20 24 17 21 27  8 24 25 25 13 23 20 17 23 21 20 24 28 27 22 21 19 23 28
#> [2857] 11 23 24 13 22 14 29 21 21 12 24 19 22 25 27 21 20 25 19 23 21 27 18 23
#> [2881] 22 23 28 20 21 23 22 23 26  9 19 15 21 18 25 20 15 21 19 19 20 23 21 23
#> [2905] 11 13 23 27 23 15 25 27 13 19 19 23 21 20 23 20 25 22 20 20 15 13 25 19
#> [2929] 17 23 25 26 24 21 25 17 26 15 21 21 18 21 23 20 26 25 15  7 25 24 23 19
#> [2953] 19 20 27 21 13 25 26 22 25 22 21 20 17 22 20 23 20 17 18 20 26 19 26 19
#> [2977] 20 20 21 19 17 14 25 18 18 13 21 15 19 13 25 27 19 16 18 25 19 18 10 26
#> 
#> $csem
#>    [1] 2.864622 2.584921 2.698484 1.870829 2.432731 2.715277 2.696799 2.735291
#>    [9] 2.473496 2.715277 2.408319 2.800433 1.954017 2.533413 2.698484 2.830034
#>   [17] 2.667992 2.480225 2.748553 2.693426 2.265686 2.877815 2.476557 2.469818
#>   [25] 2.560185 2.432731 2.334848 2.718065 2.291618 2.000000 2.656039 2.630244
#>   [33] 2.602446 2.152518 2.877815 2.533413 2.883600 2.299539 2.877815 2.540580
#>   [41] 2.800433 2.073644 2.784698 2.572583 2.533413 2.432731 2.024846 2.518417
#>   [49] 2.771609 1.957890 2.473496 2.572583 2.656039 2.893567 2.865151 2.152518
#>   [57] 2.776525 2.848711 1.897367 2.865151 1.683251 2.789591 1.612452 2.865151
#>   [65] 2.630244 2.598368 2.751859 2.148996 2.721965 2.258318 2.830034 2.718065
#>   [73] 2.435843 2.598951 2.667992 2.265686 1.957890 2.452581 2.426495 2.469818
#>   [81] 2.236068 2.771609 2.572583 2.598368 2.835917 2.754610 2.306118 2.432731
#>   [89] 2.865151 2.771609 2.883600 2.718065 2.630244 2.691175 2.144761 2.883600
#>   [97] 2.366432 2.848711 2.667992 2.830034 2.426495 1.852926 2.721965 2.789591
#>  [105] 2.195036 1.732051 2.693426 2.473496 2.693426 2.883600 2.269695 2.236068
#>  [113] 2.258318 2.800433 2.787961 2.630244 1.683251 1.870829 2.877815 2.469818
#>  [121] 2.486326 2.265686 2.432731 2.718065 2.647469 2.469818 2.396336 2.628515
#>  [129] 2.435843 2.848711 2.435843 1.414214 1.897367 2.236068 2.830034 2.789591
#>  [137] 2.584921 1.414214 2.366432 2.258318 2.291618 2.787961 2.598368 2.445776
#>  [145] 2.299539 2.110579 1.848833 1.870829 2.540580 2.467363 2.696799 2.848711
#>  [153] 1.414214 2.081666 1.000000 2.679891 2.748553 2.148996 1.906130 2.533413
#>  [161] 2.258318 2.628515 2.476557 1.683251 2.841255 2.748553 2.835917 2.647469
#>  [169] 2.637147 2.236068 2.848711 2.598368 2.848711 2.584921 2.309401 2.105548
#>  [177] 2.000000 2.740272 2.073644 2.679891 2.105548 2.533413 2.440194 2.121320
#>  [185] 1.897367 2.258318 2.776525 2.334848 2.412091 2.754610 2.572583 2.432731
#>  [193] 2.835917 2.656039 2.877815 2.656039 2.473496 2.789591 2.190890 1.852926
#>  [201] 2.509980 2.718065 2.883600 2.647469 2.598368 2.473496 2.771609 2.302173
#>  [209] 2.718065 2.877815 2.432731 2.540580 2.698484 2.432731 2.440194 2.572583
#>  [217] 1.870829 2.630244 2.584921 1.954017 2.605356 1.910894 2.656039 2.748553
#>  [225] 2.835917 2.691175 1.414214 2.258318 2.078024 2.656039 1.732051 2.899321
#>  [233] 2.784698 2.156737 2.152518 2.776525 2.776525 2.533413 2.893567 2.405801
#>  [241] 2.432731 2.636573 2.835917 2.598368 2.531618 2.572583 2.748553 2.636573
#>  [249] 2.598951 2.335497 2.408319 2.306118 2.426495 2.630244 2.598368 2.883600
#>  [257] 2.656039 2.513599 2.440194 2.912356 2.110579 2.893567 2.291618 2.467363
#>  [265] 2.540580 2.718065 2.302173 1.000000 2.698484 1.910894 2.498484 1.732051
#>  [273] 2.560185 2.408319 2.366432 2.647469 2.637147 1.852926 2.830034 1.612452
#>  [281] 2.741378 2.877815 2.718065 1.870829 2.426495 2.430862 2.366432 2.718065
#>  [289] 2.771609 2.584921 2.628515 2.748553 2.408319 2.540580 2.258318 2.748553
#>  [297] 2.740272 2.105548 1.949359 2.647469 2.691175 2.533413 2.776525 2.258318
#>  [305] 2.299539 2.513599 2.598368 2.865151 2.610585 2.156737 2.647469 2.518417
#>  [313] 2.912356 2.776525 2.584921 2.800433 1.414214 2.584921 2.078024 2.877815
#>  [321] 2.258318 2.195036 1.910894 2.000000 2.748553 2.691175 2.630244 2.800433
#>  [329] 2.156737 2.883600 2.291618 1.612452 2.412091 2.473496 2.533413 2.432731
#>  [337] 2.533413 2.598368 1.870829 2.912356 1.902152 2.248232 2.848711 2.605356
#>  [345] 2.883600 2.334848 2.081666 2.531618 1.852926 2.748553 2.408319 2.571994
#>  [353] 1.954017 2.509980 2.498484 2.000000 2.480225 2.691175 2.830034 2.024846
#>  [361] 2.784698 2.078024 2.531618 2.347145 2.776525 2.771609 2.156737 2.800433
#>  [369] 2.299539 2.258318 2.718065 2.430862 2.432731 2.473496 2.557816 2.105548
#>  [377] 2.771609 1.858641 2.667992 2.334848 2.432731 2.531618 2.306118 2.571994
#>  [385] 2.679891 2.741378 2.105548 2.877815 2.518417 1.957890 2.105548 0.000000
#>  [393] 1.566699 2.152518 2.912356 2.800433 2.912356 2.835917 1.414214 2.605356
#>  [401] 2.476557 2.848711 2.177711 2.412091 2.432731 2.718065 2.656039 1.897367
#>  [409] 2.533413 2.877815 2.467363 2.637147 2.105548 2.696799 2.899321 2.825211
#>  [417] 2.469818 2.366432 2.637147 2.864622 2.605356 2.691175 2.912356 2.469818
#>  [425] 1.683251 2.258318 1.852926 2.735291 1.957890 2.800433 2.841255 2.432731
#>  [433] 2.883600 2.830034 2.789591 2.877815 1.264911 2.101947 2.771609 2.598368
#>  [441] 1.910894 2.877815 2.605356 2.412091 2.865151 2.347145 2.841255 2.215647
#>  [449] 2.630244 2.800433 1.910894 2.741378 2.656039 2.366432 2.000000 2.101947
#>  [457] 2.598951 2.776525 2.560185 2.404541 2.152518 1.000000 2.426495 2.748553
#>  [465] 2.848711 2.877815 2.771609 2.325745 1.910894 1.910894 2.757909 2.435843
#>  [473] 2.605356 2.679891 2.721965 2.605356 2.740272 2.560185 2.605356 2.432731
#>  [481] 2.754610 2.291618 2.602446 2.473496 2.345208 2.584921 2.236068 2.715277
#>  [489] 2.432731 2.800433 2.560185 1.683251 2.408319 2.258318 2.883600 2.236068
#>  [497] 2.830034 2.476557 2.748553 2.518417 2.412091 1.897367 2.741378 2.445776
#>  [505] 2.070719 2.757909 2.835917 2.718065 1.678744 2.467363 2.435843 2.101947
#>  [513] 2.598951 2.715277 2.848711 2.469818 2.883600 2.741378 2.533413 2.800433
#>  [521] 2.473496 2.656039 2.754610 2.408319 2.533413 2.825211 2.571994 2.800433
#>  [529] 2.656039 2.258318 2.698484 2.396336 2.426495 2.335497 2.787961 2.912356
#>  [537] 2.571994 2.776525 2.335497 2.121320 2.835917 2.325745 2.432731 2.748553
#>  [545] 2.912356 2.740272 2.044949 1.957890 2.741378 2.848711 2.258318 2.258318
#>  [553] 2.784698 2.572583 2.531618 2.291618 2.912356 2.073644 1.949359 2.784698
#>  [561] 2.105548 2.927922 2.789591 2.504541 2.432731 2.156737 2.509980 2.877815
#>  [569] 2.291618 2.557816 2.345208 2.800433 2.533413 2.748553 2.518417 2.024846
#>  [577] 2.366432 2.258318 1.870829 2.473496 2.509980 2.825211 2.572583 2.848711
#>  [585] 2.152518 2.473496 2.679891 2.598368 2.302173 2.121320 2.667992 2.696799
#>  [593] 2.598951 2.440194 2.156737 2.830034 2.571994 2.698484 2.110579 2.533413
#>  [601] 2.144761 2.771609 2.412091 2.156737 2.630244 2.841255 2.771609 2.024846
#>  [609] 2.841255 2.152518 2.101947 2.105548 2.805838 2.426495 2.552479 2.718065
#>  [617] 2.656039 2.334848 2.899321 2.513599 2.740272 2.899321 2.306118 2.073644
#>  [625] 2.144761 2.152518 1.612452 2.073644 1.414214 2.473496 2.435843 2.105548
#>  [633] 2.509980 2.740272 2.408319 1.870829 2.366432 2.105548 2.302173 2.412091
#>  [641] 2.269695 1.414214 2.306118 1.897367 1.897367 2.748553 2.473496 2.691175
#>  [649] 2.306118 1.949359 2.754610 2.258318 2.637147 2.476557 2.291618 2.718065
#>  [657] 2.637147 2.258318 2.258318 2.805838 1.852926 2.156737 2.721965 2.110579
#>  [665] 2.848711 2.848711 2.605356 2.848711 1.870829 2.408319 2.691175 1.581139
#>  [673] 2.121320 2.835917 2.741378 2.334848 1.957890 2.258318 2.741378 2.748553
#>  [681] 2.740272 2.848711 2.656039 2.883600 2.148996 2.121320 1.000000 2.667992
#>  [689] 2.771609 2.893567 2.518417 2.830034 2.258318 1.897367 2.408319 2.269695
#>  [697] 2.480225 2.383656 2.432731 2.412091 2.754610 2.800433 2.572583 2.830034
#>  [705] 2.787961 2.540580 2.912356 2.696799 2.291618 1.897367 2.531618 2.630244
#>  [713] 2.776525 2.584921 2.334848 2.572583 2.748553 2.302173 2.152518 2.835917
#>  [721] 2.848711 2.426495 2.221111 2.347145 2.740272 2.754610 2.105548 2.412091
#>  [729] 2.656039 1.870829 1.906130 2.789591 2.533413 2.480225 2.291618 2.291618
#>  [737] 2.691175 2.540580 2.598368 2.190890 2.513599 2.291618 2.776525 2.480225
#>  [745] 2.299539 2.408319 2.883600 2.740272 2.771609 2.776525 2.800433 2.636573
#>  [753] 2.405801 2.533413 2.883600 2.718065 2.748553 2.225881 2.533413 2.412091
#>  [761] 2.656039 2.570226 2.432731 2.748553 2.912356 2.480225 2.426495 2.531618
#>  [769] 2.825211 2.078024 2.473496 2.518417 2.656039 2.291618 2.258318 2.412091
#>  [777] 2.636573 2.771609 2.552479 2.476557 2.325745 2.531618 2.864622 2.533413
#>  [785] 2.865151 2.893567 2.121320 2.741378 2.656039 2.718065 1.957890 2.787961
#>  [793] 2.805838 2.335497 2.531618 2.912356 2.584921 2.848711 2.572583 1.683251
#>  [801] 2.221111 2.693426 2.610585 2.835917 2.560185 2.800433 2.865151 2.024846
#>  [809] 1.732051 2.533413 2.740272 2.605356 2.509980 2.787961 2.152518 2.000000
#>  [817] 2.480225 2.473496 2.787961 2.757909 2.426495 2.258318 2.452581 2.248232
#>  [825] 1.732051 2.265686 2.440194 2.656039 2.656039 2.412091 2.789591 2.864622
#>  [833] 2.598368 2.927922 2.789591 1.870829 2.598368 2.830034 2.883600 1.957890
#>  [841] 2.432731 2.835917 2.469818 2.754610 2.656039 2.721965 2.636573 2.560185
#>  [849] 2.105548 2.412091 2.412091 2.291618 2.404541 2.691175 2.144761 2.073644
#>  [857] 2.306118 2.302173 2.435843 2.848711 1.897367 2.345208 2.440194 2.426495
#>  [865] 2.291618 1.612452 2.408319 2.698484 2.432731 2.291618 2.440194 2.533413
#>  [873] 2.598368 2.740272 2.426495 2.480225 1.354006 2.156737 2.848711 1.897367
#>  [881] 2.152518 2.105548 1.683251 2.789591 2.789591 2.480225 2.698484 2.748553
#>  [889] 2.684975 2.771609 2.628515 2.789591 2.899321 2.480225 2.570226 2.148996
#>  [897] 1.264911 2.740272 2.412091 2.584921 2.684975 2.771609 2.221111 2.560185
#>  [905] 2.912356 2.366432 2.899321 2.572583 2.236068 2.156737 2.024846 2.628515
#>  [913] 2.552479 2.721965 2.776525 2.215647 2.105548 2.598951 2.784698 2.152518
#>  [921] 2.630244 2.656039 2.584921 2.306118 2.302173 1.000000 2.533413 2.148996
#>  [929] 2.152518 2.800433 2.630244 2.721965 2.258318 2.718065 2.598951 2.721965
#>  [937] 2.480225 2.291618 1.732051 2.647469 2.302173 1.897367 2.784698 1.683251
#>  [945] 2.291618 2.557816 1.954017 2.693426 2.825211 2.698484 2.830034 1.897367
#>  [953] 2.473496 2.893567 2.486326 1.732051 2.225881 1.581139 2.148996 2.081666
#>  [961] 2.715277 2.830034 2.432731 2.789591 2.637147 2.258318 2.404541 2.423371
#>  [969] 2.408319 2.432731 2.848711 2.800433 2.101947 2.476557 1.870829 2.598368
#>  [977] 2.912356 2.334848 2.531618 2.334848 2.070719 2.258318 2.426495 2.121320
#>  [985] 2.693426 2.598368 2.258318 2.912356 2.469818 2.748553 2.864622 2.748553
#>  [993] 2.696799 2.435843 1.957890 2.306118 2.432731 2.557816 1.949359 2.776525
#> [1001] 2.718065 2.776525 2.405801 2.789591 2.784698 2.467363 2.101947 1.957890
#> [1009] 2.440194 2.105548 2.748553 2.440194 2.299539 2.105548 2.366432 2.152518
#> [1017] 1.683251 2.121320 2.440194 2.325745 2.598951 2.636573 2.000000 2.848711
#> [1025] 2.560185 2.105548 2.800433 2.480225 2.221111 2.292280 2.598368 1.870829
#> [1033] 2.291618 2.430862 2.584921 2.432731 2.684975 2.805838 2.432731 2.560185
#> [1041] 2.584921 2.408319 2.636573 2.784698 2.110579 2.306118 2.306118 2.105548
#> [1049] 2.473496 1.852926 2.748553 2.598368 2.105548 1.683251 2.432731 2.789591
#> [1057] 2.291618 2.195036 2.877815 2.024846 2.656039 2.572583 2.740272 2.248232
#> [1065] 2.656039 2.105548 2.258318 2.602446 2.560185 2.432731 2.291618 1.673320
#> [1073] 2.426495 1.870829 2.248232 2.432731 2.784698 2.630244 2.800433 2.383656
#> [1081] 2.598368 2.912356 2.221111 2.078024 2.291618 2.584921 2.877815 2.647469
#> [1089] 2.476557 2.366432 2.110579 2.334848 2.637147 2.830034 1.897367 2.518417
#> [1097] 2.473496 2.656039 2.718065 2.718065 2.309401 2.258318 2.784698 2.610585
#> [1105] 2.404541 2.335497 1.852926 2.572583 2.848711 2.258318 2.486326 2.721965
#> [1113] 2.518417 2.309401 1.673320 2.693426 2.754610 2.302173 2.540580 2.584921
#> [1121] 2.469818 2.299539 2.748553 2.366432 2.771609 2.518417 2.656039 2.754610
#> [1129] 2.078024 2.221111 2.572583 1.870829 2.628515 2.647469 2.718065 2.636573
#> [1137] 1.683251 2.748553 2.259659 2.435843 1.954017 2.121320 2.656039 2.912356
#> [1145] 2.691175 2.366432 2.366432 2.480225 2.531618 2.776525 2.877815 2.693426
#> [1153] 2.636573 2.467363 2.693426 2.877815 2.584921 2.789591 1.000000 2.776525
#> [1161] 2.258318 2.698484 2.504541 2.748553 2.656039 2.452581 2.667992 2.225881
#> [1169] 1.870829 1.683251 2.306118 2.912356 2.667992 2.396336 2.471044 2.679891
#> [1177] 2.830034 2.432731 2.156737 2.835917 2.776525 2.432731 2.152518 2.789591
#> [1185] 2.152518 2.748553 2.667992 1.732051 1.612452 2.841255 2.656039 2.740272
#> [1193] 2.865151 2.404541 2.825211 2.152518 2.740272 2.865151 2.899321 2.473496
#> [1201] 2.647469 2.366432 2.572583 2.432731 2.533413 1.870829 2.637147 2.531618
#> [1209] 2.540580 2.691175 1.000000 2.121320 2.572583 2.636573 2.334848 2.335497
#> [1217] 2.776525 2.440194 2.830034 2.740272 2.787961 1.870829 2.405801 2.800433
#> [1225] 2.721965 1.870829 2.426495 2.531618 2.691175 2.628515 2.927922 2.830034
#> [1233] 2.291618 2.776525 1.414214 1.683251 2.101947 2.805838 2.110579 2.190890
#> [1241] 2.864622 2.848711 2.740272 2.865151 2.105548 2.893567 2.830034 2.800433
#> [1249] 2.148996 2.435843 2.306118 2.572583 2.572583 2.848711 2.473496 2.800433
#> [1257] 2.412091 2.830034 2.156737 2.473496 2.899321 2.121320 2.754610 2.800433
#> [1265] 2.560185 2.258318 2.721965 2.628515 2.893567 2.560185 2.081666 2.469818
#> [1273] 2.366432 1.949359 2.572583 2.101947 2.771609 2.105548 2.789591 2.498484
#> [1281] 2.848711 2.598368 2.789591 2.299539 2.078024 2.589606 2.899321 2.864622
#> [1289] 2.531618 1.673320 2.432731 2.735291 2.509980 2.570226 2.691175 1.897367
#> [1297] 2.221111 2.560185 2.105548 2.740272 2.602446 2.265686 2.698484 2.877815
#> [1305] 2.440194 2.789591 2.893567 2.073644 2.248232 2.309401 2.735291 2.691175
#> [1313] 2.865151 2.473496 2.215647 1.000000 2.877815 2.865151 2.584921 2.691175
#> [1321] 2.912356 2.771609 1.957890 0.000000 1.870829 2.830034 2.013590 2.070719
#> [1329] 2.757909 2.078024 2.865151 2.467363 2.572583 2.533413 2.584921 2.628515
#> [1337] 2.473496 2.239453 2.698484 2.830034 2.540580 2.469818 2.302173 2.345208
#> [1345] 2.440194 2.789591 1.852926 1.732051 2.784698 2.776525 2.864622 2.718065
#> [1353] 2.912356 2.789591 2.784698 2.741378 2.883600 1.848833 2.598951 2.912356
#> [1361] 1.870829 2.236068 2.195036 2.741378 2.789591 2.426495 2.693426 2.647469
#> [1369] 2.473496 2.740272 2.306118 2.877815 1.612452 2.469818 2.432731 2.691175
#> [1377] 2.789591 2.628515 1.000000 2.667992 2.560185 2.121320 2.473496 2.572583
#> [1385] 2.533413 2.531618 2.848711 2.070719 2.865151 1.910894 2.291618 2.877815
#> [1393] 2.776525 2.152518 1.354006 2.748553 2.598951 2.306118 2.789591 2.531618
#> [1401] 2.605356 2.771609 2.299539 2.440194 2.258318 2.498484 2.912356 2.684975
#> [1409] 2.637147 2.741378 2.345208 2.258318 2.754610 2.408319 2.630244 2.121320
#> [1417] 1.957890 2.533413 2.776525 2.334848 2.291618 1.852926 2.531618 2.024846
#> [1425] 2.630244 1.683251 2.605356 2.000000 2.306118 2.552479 2.787961 2.533413
#> [1433] 2.605356 2.605356 2.432731 2.121320 2.291618 2.248232 2.476557 2.432731
#> [1441] 2.656039 1.354006 2.835917 2.771609 2.540580 2.292280 2.366432 2.696799
#> [1449] 1.683251 2.800433 2.789591 2.787961 2.258318 2.740272 2.265686 2.258318
#> [1457] 2.656039 2.748553 2.841255 2.805838 2.476557 2.679891 1.732051 2.258318
#> [1465] 2.740272 2.533413 2.877815 2.848711 2.366432 1.906130 2.426495 2.718065
#> [1473] 2.110579 2.598951 2.144761 2.656039 2.784698 2.754610 1.870829 2.741378
#> [1481] 2.830034 2.513599 2.464290 2.789591 2.718065 2.259659 1.957890 2.630244
#> [1489] 2.848711 2.598951 2.741378 2.771609 2.656039 1.957890 2.740272 2.830034
#> [1497] 2.830034 2.291618 2.698484 2.121320 2.476557 2.540580 2.789591 2.771609
#> [1505] 1.414214 2.152518 2.533413 2.848711 2.572583 2.347145 2.718065 1.683251
#> [1513] 2.000000 2.504541 2.927922 2.830034 2.848711 2.531618 2.121320 2.473496
#> [1521] 2.306118 1.683251 2.225881 2.584921 2.698484 2.366432 2.848711 2.835917
#> [1529] 2.835917 2.101947 2.078024 2.883600 2.073644 2.628515 2.405801 2.435843
#> [1537] 2.432731 2.291618 2.628515 2.533413 2.334848 2.156737 2.693426 1.000000
#> [1545] 2.800433 2.156737 2.101947 2.784698 1.897367 2.718065 2.776525 2.893567
#> [1553] 2.325745 2.830034 2.152518 2.605356 2.584921 2.598368 2.513599 2.540580
#> [1561] 2.893567 2.302173 2.335497 2.693426 2.540580 2.518417 2.715277 2.345208
#> [1569] 2.636573 2.476557 2.345208 2.602446 2.877815 2.467363 2.306118 2.628515
#> [1577] 2.426495 2.835917 1.957890 2.883600 2.784698 2.805838 2.309401 2.693426
#> [1585] 2.877815 2.647469 2.598368 2.630244 2.656039 2.152518 2.144761 2.721965
#> [1593] 2.405801 2.513599 2.302173 2.835917 2.740272 2.430862 2.560185 2.105548
#> [1601] 2.718065 2.789591 2.776525 2.698484 2.334848 2.771609 2.830034 2.605356
#> [1609] 2.299539 2.504541 2.306118 2.258318 2.679891 2.383656 1.414214 2.412091
#> [1617] 2.540580 2.598951 2.509980 2.291618 2.306118 2.679891 1.612452 2.347145
#> [1625] 2.156737 2.800433 2.784698 2.560185 2.560185 2.121320 2.865151 2.805838
#> [1633] 1.414214 1.414214 2.698484 2.835917 2.221111 2.789591 2.432731 1.949359
#> [1641] 2.698484 2.848711 2.101947 2.784698 2.440194 2.258318 2.784698 2.830034
#> [1649] 2.693426 2.721965 2.435843 2.848711 2.024846 2.121320 2.518417 2.805838
#> [1657] 2.718065 2.748553 2.440194 2.647469 1.683251 2.927922 2.423371 2.078024
#> [1665] 1.683251 2.396336 2.647469 1.354006 1.954017 2.473496 1.957890 2.557816
#> [1673] 2.877815 2.533413 2.693426 2.560185 1.348400 2.306118 2.848711 2.835917
#> [1681] 2.408319 2.691175 2.412091 2.721965 1.000000 2.513599 2.605356 1.906130
#> [1689] 2.912356 2.757909 2.598951 1.858641 2.684975 2.715277 2.292280 1.897367
#> [1697] 2.469818 2.432731 2.741378 2.533413 2.696799 2.636573 2.628515 2.718065
#> [1705] 2.656039 1.957890 2.656039 2.883600 2.883600 2.518417 2.509980 2.787961
#> [1713] 2.684975 2.656039 2.248232 2.877815 2.423371 2.893567 2.693426 2.432731
#> [1721] 2.741378 2.531618 2.432731 2.721965 2.236068 2.513599 2.552479 2.152518
#> [1729] 2.473496 2.334848 2.630244 1.852926 2.435843 2.024846 2.291618 2.877815
#> [1737] 2.156737 2.121320 2.647469 2.757909 2.830034 2.432731 2.440194 2.598951
#> [1745] 2.718065 2.480225 2.473496 2.848711 2.306118 2.560185 2.000000 2.835917
#> [1753] 2.498484 2.408319 1.414214 2.540580 2.302173 1.910894 2.486326 2.637147
#> [1761] 2.605356 2.848711 2.306118 2.757909 2.679891 2.656039 2.309401 2.776525
#> [1769] 1.000000 1.354006 2.877815 2.258318 2.584921 2.073644 2.598368 2.864622
#> [1777] 2.912356 2.248232 1.870829 2.835917 1.957890 2.912356 2.598951 2.070719
#> [1785] 1.683251 2.830034 2.105548 2.533413 2.598368 2.800433 2.540580 2.560185
#> [1793] 2.883600 2.741378 2.835917 2.248232 2.412091 2.698484 2.476557 0.000000
#> [1801] 1.910894 2.877815 2.366432 2.302173 1.683251 2.467363 2.291618 2.473496
#> [1809] 2.584921 2.073644 1.581139 2.560185 2.757909 2.531618 2.584921 2.636573
#> [1817] 2.473496 2.628515 2.101947 2.630244 2.771609 2.748553 2.144761 2.000000
#> [1825] 2.584921 2.754610 2.105548 2.156737 2.584921 2.121320 2.912356 2.432731
#> [1833] 2.748553 2.258318 2.776525 2.540580 2.825211 2.789591 2.647469 2.656039
#> [1841] 2.435843 2.073644 2.570226 2.805838 2.121320 2.533413 2.302173 1.683251
#> [1849] 2.498484 2.602446 2.334848 2.757909 2.877815 2.426495 2.248232 1.732051
#> [1857] 2.557816 1.678744 2.024846 2.325745 2.787961 2.718065 2.435843 2.630244
#> [1865] 2.426495 2.893567 2.656039 2.584921 2.299539 2.637147 2.893567 2.533413
#> [1873] 1.897367 2.757909 2.748553 2.748553 2.408319 2.432731 2.693426 2.691175
#> [1881] 2.754610 2.291618 2.432731 2.291618 2.830034 2.440194 2.467363 2.518417
#> [1889] 2.306118 1.852926 2.912356 2.156737 2.432731 2.691175 2.912356 2.825211
#> [1897] 2.754610 2.248232 2.627938 1.852926 2.073644 2.628515 2.698484 2.334848
#> [1905] 2.152518 2.309401 2.656039 2.805838 2.848711 2.718065 2.269695 2.299539
#> [1913] 2.309401 2.258318 1.414214 2.721965 2.334848 2.156737 2.830034 1.678744
#> [1921] 2.540580 2.883600 2.389370 2.105548 1.732051 2.771609 2.679891 2.741378
#> [1929] 2.440194 2.787961 2.693426 2.572583 1.954017 2.877815 2.776525 1.678744
#> [1937] 2.800433 2.533413 1.897367 2.572583 2.078024 2.078024 2.259659 2.430862
#> [1945] 2.533413 2.864622 2.533413 1.957890 2.121320 2.718065 2.426495 1.683251
#> [1953] 2.656039 2.805838 2.877815 2.408319 2.432731 2.533413 2.656039 2.718065
#> [1961] 2.024846 2.152518 2.740272 2.426495 2.258318 2.637147 2.121320 2.848711
#> [1969] 2.572583 2.848711 2.560185 2.637147 2.830034 2.572583 2.432731 2.602446
#> [1977] 2.560185 2.696799 2.865151 2.598951 2.602446 2.572583 2.835917 1.683251
#> [1985] 2.572583 2.396336 2.718065 1.673320 2.513599 2.504541 2.830034 2.740272
#> [1993] 2.912356 2.718065 2.309401 2.572583 2.656039 2.605356 1.683251 2.156737
#> [2001] 2.684975 2.121320 2.432731 2.408319 2.518417 2.598368 2.784698 2.877815
#> [2009] 2.440194 2.121320 2.740272 2.190890 2.656039 2.800433 2.435843 2.602446
#> [2017] 2.471044 2.121320 2.757909 2.667992 2.698484 2.789591 2.024846 2.598368
#> [2025] 2.776525 2.865151 1.954017 1.957890 1.870829 1.732051 2.258318 2.865151
#> [2033] 1.732051 2.605356 2.647469 2.347145 2.598368 2.024846 2.698484 2.560185
#> [2041] 2.800433 1.581139 2.636573 2.513599 2.366432 2.269695 2.679891 2.865151
#> [2049] 1.870829 2.835917 2.602446 2.800433 2.540580 2.480225 2.291618 2.258318
#> [2057] 1.612452 1.858641 2.509980 2.784698 2.299539 2.105548 2.345208 1.906130
#> [2065] 2.667992 2.467363 2.389370 2.440194 2.236068 2.432731 2.412091 1.414214
#> [2073] 2.848711 2.306118 1.870829 2.835917 2.190890 2.784698 2.432731 2.667992
#> [2081] 2.637147 2.412091 2.679891 2.598368 2.236068 2.469818 2.721965 2.877815
#> [2089] 2.269695 2.748553 2.735291 2.467363 2.912356 2.841255 2.784698 2.912356
#> [2097] 2.560185 2.383656 2.748553 2.334848 2.258318 2.221111 2.877815 2.825211
#> [2105] 1.902152 2.101947 2.432731 2.691175 2.024846 2.754610 2.469818 2.684975
#> [2113] 2.602446 2.789591 2.334848 2.776525 2.800433 2.598368 2.396336 2.893567
#> [2121] 2.740272 2.291618 2.105548 2.656039 1.683251 2.912356 2.101947 2.754610
#> [2129] 2.584921 2.291618 2.265686 2.081666 1.683251 2.105548 2.572583 2.647469
#> [2137] 2.000000 2.426495 2.748553 2.771609 2.598951 2.531618 2.296242 2.557816
#> [2145] 2.825211 2.105548 2.805838 2.408319 2.835917 2.531618 2.754610 2.825211
#> [2153] 2.560185 2.754610 2.513599 2.877815 2.848711 2.805838 1.957890 2.473496
#> [2161] 2.656039 2.800433 2.412091 2.877815 2.754610 2.771609 2.236068 2.754610
#> [2169] 2.396336 2.628515 2.696799 2.877815 2.721965 2.480225 2.693426 2.073644
#> [2177] 2.679891 2.540580 1.870829 2.073644 2.412091 2.366432 1.732051 2.877815
#> [2185] 2.306118 2.408319 2.883600 2.715277 1.732051 2.628515 2.152518 2.877815
#> [2193] 2.598368 2.366432 2.432731 2.598951 1.341641 2.584921 2.693426 2.540580
#> [2201] 2.789591 2.754610 2.691175 2.366432 2.024846 2.121320 2.698484 2.435843
#> [2209] 2.000000 2.877815 2.258318 2.258318 2.306118 1.852926 2.258318 2.432731
#> [2217] 2.584921 2.741378 2.698484 2.865151 2.656039 2.291618 2.637147 2.432731
#> [2225] 2.533413 2.560185 2.771609 1.957890 2.480225 2.432731 2.927922 2.740272
#> [2233] 2.302173 2.647469 2.628515 2.306118 2.835917 2.531618 2.656039 2.572583
#> [2241] 2.476557 2.480225 2.572583 2.540580 2.412091 2.345208 0.000000 1.852926
#> [2249] 1.870829 2.636573 2.024846 1.581139 2.073644 2.647469 2.741378 2.754610
#> [2257] 1.870829 2.152518 2.691175 2.787961 2.757909 2.467363 2.473496 2.435843
#> [2265] 2.667992 2.156737 2.740272 1.414214 2.598951 2.877815 2.402650 2.630244
#> [2273] 2.757909 1.414214 2.473496 2.557816 2.552479 2.000000 2.366432 2.598368
#> [2281] 2.893567 2.302173 2.787961 2.473496 2.347145 2.013590 2.291618 2.258318
#> [2289] 2.647469 2.440194 2.440194 2.787961 2.078024 2.864622 2.292280 2.148996
#> [2297] 2.572583 1.897367 2.299539 2.473496 2.533413 1.870829 2.513599 2.440194
#> [2305] 2.630244 2.656039 1.954017 1.910894 2.412091 2.557816 2.748553 2.636573
#> [2313] 1.341641 2.805838 2.572583 2.467363 1.678744 2.628515 2.789591 2.605356
#> [2321] 2.335497 1.683251 2.718065 2.426495 2.258318 2.740272 2.740272 1.870829
#> [2329] 2.754610 2.636573 2.236068 2.177711 2.024846 2.292280 2.258318 1.414214
#> [2337] 2.721965 2.830034 2.366432 2.408319 2.927922 2.691175 2.698484 2.848711
#> [2345] 2.195036 2.754610 2.825211 2.435843 2.258318 1.673320 2.073644 2.748553
#> [2353] 2.789591 2.605356 2.656039 2.469818 1.000000 2.236068 1.683251 2.473496
#> [2361] 2.000000 2.698484 2.740272 2.302173 2.476557 2.754610 2.306118 2.684975
#> [2369] 2.426495 2.572583 2.309401 2.498484 2.691175 2.598951 2.486326 2.557816
#> [2377] 2.830034 2.899321 2.584921 2.258318 2.679891 1.870829 2.302173 2.306118
#> [2385] 2.656039 2.693426 2.440194 2.899321 2.877815 2.306118 2.078024 2.334848
#> [2393] 2.078024 2.656039 2.560185 1.683251 2.572583 2.848711 1.848833 2.848711
#> [2401] 2.693426 2.656039 2.877815 2.602446 2.533413 2.299539 2.864622 1.852926
#> [2409] 2.533413 2.572583 1.354006 1.870829 1.414214 2.865151 2.533413 2.533413
#> [2417] 2.602446 2.718065 2.531618 2.531618 2.656039 2.771609 2.476557 2.721965
#> [2425] 2.081666 2.656039 2.825211 2.656039 2.784698 2.584921 2.190890 1.852926
#> [2433] 2.912356 2.258318 1.870829 2.787961 2.498484 2.776525 2.533413 1.954017
#> [2441] 2.467363 2.848711 2.366432 2.912356 2.473496 2.757909 2.647469 2.432731
#> [2449] 2.152518 2.557816 2.258318 2.306118 2.476557 2.864622 2.893567 1.957890
#> [2457] 2.841255 2.789591 1.852926 2.291618 2.628515 2.864622 2.584921 2.412091
#> [2465] 2.440194 1.852926 1.906130 2.748553 2.630244 1.354006 2.473496 2.721965
#> [2473] 2.718065 2.636573 2.309401 2.105548 2.432731 2.835917 2.291618 2.864622
#> [2481] 2.877815 2.647469 2.540580 2.899321 2.572583 2.647469 2.800433 1.683251
#> [2489] 2.598951 2.835917 2.412091 2.748553 2.598368 2.877815 2.334848 2.533413
#> [2497] 2.584921 2.757909 2.877815 2.518417 2.864622 2.334848 2.452581 2.628515
#> [2505] 2.656039 2.221111 2.552479 2.628515 1.732051 2.571994 2.718065 2.893567
#> [2513] 2.412091 2.912356 2.405801 2.598368 2.432731 2.771609 2.513599 2.637147
#> [2521] 2.865151 2.865151 2.440194 2.347145 2.000000 2.078024 2.335497 2.647469
#> [2529] 1.683251 2.024846 2.306118 2.800433 2.513599 2.864622 2.698484 1.612452
#> [2537] 2.024846 2.679891 2.221111 2.605356 2.396336 2.693426 2.805838 2.598368
#> [2545] 1.954017 2.656039 1.683251 1.957890 2.748553 2.825211 2.469818 2.830034
#> [2553] 2.473496 2.513599 2.467363 2.684975 2.101947 2.440194 1.732051 2.693426
#> [2561] 2.366432 1.612452 1.870829 2.841255 2.291618 2.366432 2.848711 2.473496
#> [2569] 2.533413 2.630244 2.754610 2.605356 2.735291 2.531618 2.258318 2.408319
#> [2577] 2.334848 1.683251 2.721965 2.741378 2.110579 2.800433 2.221111 2.408319
#> [2585] 2.698484 2.718065 2.156737 2.426495 2.513599 2.721965 1.870829 2.830034
#> [2593] 2.741378 2.787961 2.628515 2.473496 2.754610 2.741378 2.291618 2.748553
#> [2601] 2.572583 1.683251 2.473496 2.325745 2.698484 2.306118 2.309401 2.101947
#> [2609] 2.435843 2.789591 2.912356 2.121320 2.044949 2.789591 1.414214 1.897367
#> [2617] 2.432731 2.789591 2.518417 2.347145 1.852926 2.101947 2.435843 2.258318
#> [2625] 2.636573 2.265686 2.408319 2.144761 2.584921 2.435843 2.718065 2.776525
#> [2633] 2.830034 2.741378 2.605356 1.949359 2.236068 2.684975 2.848711 2.656039
#> [2641] 2.787961 2.225881 2.440194 2.800433 2.533413 2.754610 1.732051 2.509980
#> [2649] 2.883600 1.949359 2.509980 2.236068 2.698484 2.156737 2.630244 2.789591
#> [2657] 2.741378 2.467363 2.533413 2.883600 2.751859 2.841255 2.899321 2.258318
#> [2665] 2.584921 2.864622 2.533413 2.754610 1.858641 1.906130 2.435843 2.291618
#> [2673] 2.101947 2.602446 2.000000 2.715277 2.721965 2.825211 2.877815 2.865151
#> [2681] 2.560185 2.696799 2.684975 2.426495 2.698484 2.334848 2.258318 2.830034
#> [2689] 2.825211 2.754610 2.258318 2.751859 1.957890 2.540580 2.557816 2.221111
#> [2697] 2.335497 2.684975 2.718065 2.877815 2.789591 2.156737 2.800433 2.630244
#> [2705] 1.957890 2.637147 2.667992 2.605356 2.195036 2.691175 2.291618 2.473496
#> [2713] 2.877815 2.432731 2.473496 2.070719 2.912356 2.877815 2.598951 2.533413
#> [2721] 2.748553 2.426495 2.883600 2.101947 2.800433 2.656039 2.864622 2.693426
#> [2729] 2.309401 2.667992 2.430862 1.732051 2.789591 2.741378 2.754610 2.893567
#> [2737] 2.000000 2.540580 2.430862 2.789591 2.560185 1.957890 1.910894 2.258318
#> [2745] 2.110579 1.870829 2.598368 2.432731 2.825211 2.291618 2.152518 2.679891
#> [2753] 2.691175 2.366432 2.771609 2.912356 2.306118 1.673320 1.957890 2.830034
#> [2761] 2.121320 2.366432 2.784698 2.292280 2.557816 2.877815 2.121320 2.598368
#> [2769] 2.533413 1.957890 2.432731 2.630244 2.696799 2.024846 2.215647 2.605356
#> [2777] 2.748553 2.383656 2.865151 2.893567 2.636573 2.789591 2.848711 2.877815
#> [2785] 2.805838 2.276361 2.258318 2.105548 2.334848 2.533413 2.912356 2.531618
#> [2793] 2.848711 2.899321 2.396336 2.748553 2.693426 2.656039 2.718065 2.656039
#> [2801] 2.469818 2.473496 2.432731 2.805838 2.073644 2.258318 2.667992 2.571994
#> [2809] 1.000000 2.771609 1.000000 2.789591 2.531618 2.598368 2.877815 2.718065
#> [2817] 2.656039 2.473496 2.602446 2.531618 2.679891 2.691175 2.531618 2.258318
#> [2825] 2.735291 2.509980 2.864622 2.584921 2.467363 2.598368 1.683251 1.957890
#> [2833] 2.412091 2.302173 2.691175 2.693426 1.910894 2.531618 2.412091 1.897367
#> [2841] 2.306118 2.789591 2.531618 2.741378 2.754610 2.533413 2.306118 2.800433
#> [2849] 2.292280 1.678744 1.957890 2.647469 2.533413 2.531618 2.144761 1.683251
#> [2857] 2.776525 2.533413 1.732051 2.864622 2.347145 2.893567 1.414214 2.679891
#> [2865] 2.693426 2.805838 2.366432 2.825211 2.598951 2.177711 1.732051 2.740272
#> [2873] 2.800433 2.299539 2.830034 2.480225 2.560185 1.957890 2.718065 2.473496
#> [2881] 2.557816 2.540580 1.581139 2.667992 2.693426 2.404541 1.683251 2.531618
#> [2889] 2.024846 2.584921 2.540580 2.679891 2.718065 2.800433 2.291618 2.667992
#> [2897] 2.877815 2.679891 2.540580 2.848711 2.480225 2.408319 2.636573 2.540580
#> [2905] 2.480225 2.628515 2.291618 1.957890 2.518417 2.771609 2.073644 1.897367
#> [2913] 2.789591 2.531618 2.830034 2.291618 2.693426 2.800433 2.408319 2.776525
#> [2921] 2.306118 2.598368 2.776525 2.667992 2.771609 2.883600 2.291618 2.848711
#> [2929] 2.771609 2.473496 2.144761 1.906130 2.309401 2.236068 2.215647 2.899321
#> [2937] 2.156737 2.877815 2.693426 2.693426 2.800433 2.509980 2.533413 2.630244
#> [2945] 2.105548 2.177711 2.893567 2.396336 2.258318 2.334848 2.560185 2.740272
#> [2953] 2.584921 2.440194 1.957890 2.630244 2.498484 2.081666 2.156737 2.598951
#> [2961] 2.299539 2.598368 2.560185 2.396336 2.830034 2.572583 2.667992 2.560185
#> [2969] 2.667992 2.877815 2.883600 2.598368 2.105548 2.835917 1.949359 2.830034
#> [2977] 2.789591 2.656039 2.740272 2.748553 2.899321 2.560185 2.152518 2.605356
#> [2985] 2.605356 2.533413 2.718065 2.552479 2.667992 2.584921 2.306118 1.870829
#> [2993] 2.848711 2.848711 2.605356 2.258318 2.667992 2.864622 2.504541 2.156737
#>