Gaussian quadrature points and weights

Generate Gaussian quadrature points and corresponding weights

normal_quadra(n, mm)

Arguments

n

Number of quadrature points

mm

Absolute value of minumum or maximum of quadrature points

Value

A list of Gaussian quadrature points and corresponding weights

Examples

normal_quadra(41, 5)
#> $nodes
#>  [1] -5.00 -4.75 -4.50 -4.25 -4.00 -3.75 -3.50 -3.25 -3.00 -2.75 -2.50 -2.25
#> [13] -2.00 -1.75 -1.50 -1.25 -1.00 -0.75 -0.50 -0.25  0.00  0.25  0.50  0.75
#> [25]  1.00  1.25  1.50  1.75  2.00  2.25  2.50  2.75  3.00  3.25  3.50  3.75
#> [37]  4.00  4.25  4.50  4.75  5.00
#> 
#> $weights
#>  [1] 3.716800e-07 1.257377e-06 3.995936e-06 1.192966e-05 3.345757e-05
#>  [6] 8.814895e-05 2.181707e-04 5.072622e-04 1.107962e-03 2.273391e-03
#> [11] 4.382076e-03 7.934915e-03 1.349775e-02 2.156934e-02 3.237941e-02
#> [16] 4.566228e-02 6.049270e-02 7.528438e-02 8.801636e-02 9.666706e-02
#> [21] 9.973560e-02 9.666706e-02 8.801636e-02 7.528438e-02 6.049270e-02
#> [26] 4.566228e-02 3.237941e-02 2.156934e-02 1.349775e-02 7.934915e-03
#> [31] 4.382076e-03 2.273391e-03 1.107962e-03 5.072622e-04 2.181707e-04
#> [36] 8.814895e-05 3.345757e-05 1.192966e-05 3.995936e-06 1.257377e-06
#> [41] 3.716800e-07
#>